backend/requirements.txt
# Core dependencies
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.9
pillow>=10.0
requests>=2.31
httpx>=0.27.0
pytest>=7.4
google-genai>=0.6.0
openai>=1.0.0
pydantic>=2.0
python-dotenv>=1.0.1
sarvamai
